swaram_backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
gunicorn==21.2.0